refactor: create a handler for setting a token expire time
create a handler for setting a token expire time after a page reloading

MODULE-1-PART-3-gallery
