+++ gallery/gallery.js
@@ -1,13 +1,13 @@
-import { setExpireTimeAfterReloading } from "../utilities/token/setToken.js";
 import { renderGalleryPage } from "/gallery/handler.js";
 import { reRenderGalleryPage } from "/gallery/handler.js"
 import { checkTokenExists } from "/gallery/handler.js";
+import { setExpireTimeAfterReloading } from "/gallery/handler.js";
 
 setExpireTimeAfterReloading();
 
-checkTokenExists(); // handler
+checkTokenExists();
 
-await renderGalleryPage(); // handler
+await renderGalleryPage();
 
 const pages = document.getElementById('pages');
-pages.onclick = await reRenderGalleryPage(); // handler
+pages.onclick = await reRenderGalleryPage();
