+++ gallery/handler.js
@@ -50,3 +50,13 @@ export function checkTokenExists() {
         urlService.redirectToLoginPage();
     }
 }
+
+export function setExpireTimeAfterReloading() {
+    const currentTime = Date.now();
+    const tokenExpireTime = localStorage.getItem('tokenExpireTime');
+    const timeLeft = tokenExpireTime - currentTime;
+
+    setTimeout(() => {
+        localStorage.removeItem('token');
+    }, timeLeft);
+}
