--- utilities/token/setToken.js
@@ -1,9 +1,0 @@
-export function setExpireTimeAfterReloading() {
-    const currentTime = Date.now();
-    const tokenExpireTime = localStorage.getItem('tokenExpireTime');
-    const timeLeft = tokenExpireTime - currentTime;
-
-    setTimeout(() => {
-        localStorage.removeItem('token');
-    }, timeLeft);
-}
